requirements.txt
google-generativeai
python-dotenv
langchain
pypdf
faiss-cpu
langchain_google_genai
langchain_community
requests
streamlit
fitz
PyMuPDF

